==> docs/eDNA 12S metab/06-phyloseq-template-12S.Rmd <==
---
title: "Phyloseq Diversity Metrics: eDNA metabarcoding base script"
output:
  github_document: default
  pdf_document:
    keep_tex: yes
  html_document:
    toc: yes
    toc_depth: 6
    toc_float: yes
editor_options: 
  chunk_output_type: inline
---

This script analyzes your relative abundance matrix to assess alpha and beta diversity. Figures produced are potentially part of the main figures of your manuscript/report. 

# Load libraries

```{r}
library(ggplot2) ## for plotting
library(tidyverse) ## for data manipulation 
library(phyloseq)
library(knitr)
library(readxl)
library(writexl)
library(cowplot)

## for stats
library(pairwiseAdonis)
library(lme4) ## for stats
library(car) ## for stats
library(stats) ## for stats
library(vegan)
library("microbiome") ## for alpha diversity functions

## set seed
set.seed(1234)
```

# Load data 

Before continuing with analyses, decide on what data you're going to use an input. Relative abundance, raw reads, rarefied counts (can be done in phyloseq), variance stabilizing transformation (vst) from DESeq2? The input will impact your interpretation and thus is important to decide **before** conducting any stats to avoid bias towards a particular result.

On the Fisheries team, we have traditionally used relative abundance and the following code uses that dataset.

### Relative Abundance data

```{r}
# Option 1: Change global options (affects all subsequent operations)
## Convert scientific notation to regular numbers
options(scipen = 999)

## Relative abundance matrix 
df <- read_xlsx("docs/eDNA 12S metab/example_output/Results_matrix_relative.xlsx") %>%
  ## removing common_name and category for now 
  dplyr::select(-Common_name, -Category) %>%
  
  ## Remove columns with NA values
  dplyr::select(where(~!any(is.na(.)))) %>%
  
  ## making species_name rownames instead of column 
  column_to_rownames(var = "Species_name") %>%
  
  # remove columns that sum to 0
  select(where(~sum(., na.rm = TRUE) != 0))
```

### Metadata

```{r}
meta <- read_xlsx("docs/eDNA 12S metab/example_input/metadata_tidal_cycle.xlsx") %>%
  
    ## match sample IDs to matrix if needed 
  mutate(`GMGI Sample ID` = gsub("-", "_", `GMGI Sample ID`),
         `GMGI Sample ID` = paste0(`GMGI Sample ID`, "_12S")) %>%
  
  ## rownames are also needed in phyloseq meta table
  mutate(sampleID2=`GMGI Sample ID`) %>% column_to_rownames(var = "sampleID2") 
```

# Create phyloseq object

```{r}
## Create ASV (OTU) table and meta table 
otu <- otu_table(df, taxa_are_rows = T)
meta_phyloseq <- sample_data(meta)

## Merge metadata and OTU table into one phyloseq "object"
phylo_obj <- merge_phyloseq(otu, meta_phyloseq)

## view phyloseq obj 
## expected output = otu_table() with taxa and sample numbers and sample_data() with the sample and column numbers
print(phylo_obj)

# Ensure that your OTU table doesn't contain any NA or negative values (output should be FALSE)
any(is.na(otu_table(phylo_obj)))
any(otu_table(phylo_obj) < 0)
```

### Subsetting phyloseq object

Optional if you'd like to break your data up by a certain variable (surface water, bottom water). If desired, uncomment the code chunk below.

```{r}
# surface <- subset_samples(phylo_obj, SampleType == "Surface Water")
# bottom <- subset_samples(phylo_obj, SampleType == "Bottom Water") 
```

# Alpha Diversity 

Comparing the species diversity (shannon index or species richness) at each site. 

## Calculating Shannon Index and Species Richness

The alpha function in the microbiome package calculates several alpha diversity indices. The most relevant are likely observed (species richness) and diversity_shannon (shannon index). 

https://microbiome.github.io/tutorials/Alphadiversity.html  

Or use plot_richness function from:  
https://rstudio-pubs-static.s3.amazonaws.com/1071936_6115f873acbc4dc4a30b1380cc3885fb.html 

```{r}
### USER: CHANGE COLORS AS NEEDED
fill_col = c("cyan4", "deeppink4", "#ed6a5a", "yellow2")

## Calculate
alpha_div <- estimate_richness(phylo_obj, measures = c("Shannon", "Simpson")) %>%
  rownames_to_column(var = "GMGI Sample ID") %>% left_join(., meta, by = "GMGI Sample ID") 

### USER: EDIT THE X AXIS AND THE FILL=, SHAPE= AESTHETICS 
alpha_div %>%
  gather("measurement", "value", Shannon:Simpson) %>%
  ggplot(., aes(x=Area, y=value)) +
  geom_boxplot(aes(color = Area), fill=NA, outlier.shape=NA, show.legend = FALSE) +
  geom_jitter(width=0.2, shape=21, aes(fill = Area), color = 'black', alpha=0.5) +
  facet_wrap(~measurement, scales = "free_y", strip.position = "left") +
  ## labels
  labs(x="Area", y="", fill = "Area") +
  
  scale_fill_manual(values = fill_col) +
  scale_color_manual(values = fill_col) +

  ## theme options
  theme_bw() + 
  theme(panel.background=element_blank(),
        strip.background=element_blank(),
        strip.text = element_text(size = 10, face="bold"),
        legend.position = "none",
        strip.clip = 'off',
        strip.placement = "outside",
        axis.text.y = element_text(size=8, color="grey30"),
        axis.text.x = element_text(size=8, color="grey30"),
        axis.title.y = element_text(margin = margin(t = 0, r = 10, b = 0, l = 0), size=10, face="bold"),
        axis.title.x = element_text(margin = margin(t = 10, r = 0, b = 0, l = 0), size=10, face="bold"))

ggsave("docs/eDNA 12S metab/example_output/figures/alpha_diversity.png", width = 6, height = 4)
```

## Statistics

Test: T-test or ANOVA 
- Type I, II, and III:   
- * indicates an interaction (SampleType*Month). Usually we are interested in the interaction of our factors.  
- + indicates an additive effect (SampleType + Month)

T-test to be used when only two groups to compare and ANOVA to be used with 3+ groups. Al

non-parametric Kolmogorov-Smirnov test for two-group comparisons when there are no relevant covariates

```{r}
## Create model 
aov <- aov(Shannon ~ Area, data = alpha_div)

## ANOVA test on above model
Anova(aov, type = "III")
```

Test: If 3+ groups, Tukey Post Hoc Comparisons 

```{r}
TukeyHSD(aov)
```


# Beta Diversity 

Comparing the community assemblages between sites/groups.

Resources (Read before continuing):  
- https://ourcodingclub.github.io/tutorials/ordination/
- https://uw.pressbooks.pub/appliedmultivariatestatistics/chapter/comparison-of-ordination-techniques/ 

Common options:  
- Principal Components Analysis (PCA): Euclidean distance measure   
- Principal Coordinates Analysis (PCoA): Dissimilarity distance based   
- Non-metric MultiDimensional Scaling (NMDS): Dissimilarity distance based  

PCoA and NMDS handle zero's in community matrices much better than PCA. For metabarcoding data, usually PCoA and NMDS are more appropriate. The differences between PCoA and NMDS are minor compared to difference between PCA. NMDS is iterative and used a different ordering method (see resource links above for more info). 

NMDS requires evaluation of the output 'stress value': This value tells you how well the model fit your data. This is helpful to include on your NMDS plot in reports/manuscripts/presentations. 

Stress (0-1 scale)	Interpretation
`< 0.05`	= Excellent representation with no prospect of misinterpretation
`< 0.10`	= Good ordination with no real disk of drawing false inferences
`< 0.20`	= Can be useful but has potential to mislead.  In particular, shouldn’t place too much reliance on the details
`> 0.20`	= Could be dangerous to interpret
`> 0.35`	= Samples placed essentially at random; little relation to original ranked distances

If your stress value is >0.2, do not include in analyses and try PCoA instead.

## Calculating dissimilarity matrix

```{r}
## Bray Curtis Dissimilarity Matrix (used in statistics)
bray_df <- phyloseq::distance(phylo_obj, method = "bray")

## Sample information
sample_df <- data.frame(sample_data(phylo_obj))
```

## Plotting 

### PCoA 

https://www.rdocumentation.org/packages/phyloseq/versions/1.16.2/topics/ordinate 

```{r}
## Conduct PCoA 
pcoa <- ordinate(physeq = phylo_obj, method = "PCoA", distance = "bray")

## Plotting
plot_ordination(phylo_obj, pcoa, 
                
                ## USER EDITS shape, color, alpha, fill, etc. as desired based on project metadata
                color = "Area") +
  
  ## Point and point aesthetics
  geom_point(aes(color = Area), alpha = .5, size = 5) +
  scale_color_manual(values = c("green4", "gold3", "pink2", "purple")) +

  ## Labels: USER EDITS as desired
  labs(color = "Area") +
  ggtitle("PCoA example") +
  
  ## Theme: USER EDITS as desired
  theme_bw() +
  theme(
    legend.title = element_text(face = "bold", size=12),
    legend.position = "right",
    axis.title.y = element_text(margin = margin(t = 0, r = 10, b = 0, l = 0), size=10, face="bold"),
    axis.title.x = element_text(margin = margin(t = 10, r = 0, b = 0, l = 0), size=10, face="bold"),
  ) 
  
## USER EDITS WIDTH AND HEIGHT TO DESIRED   
ggsave("docs/eDNA 12S metab/example_output/figures/PCoA_phyloseq.png", width = 8, height = 6)
```

### NMDS

```{r}
## Calculating NMDS 
## filtering example if needed
# phylo_obj_filtered <- prune_samples(sample_data(phylo_obj)$SampleType != "Blank", phylo_obj)

NMDS <- ordinate(physeq = phylo_obj, method = "NMDS", distance = "bray")

## Plotting
plot_ordination(phylo_obj, NMDS, 
                
                ## USER EDITS shape, color, alpha, fill, etc. as desired based on project metadata
                color = "Area") +
  
  ## Point and point aesthetics
  geom_point(aes(color = Area), alpha = .5, size = 5) +
  scale_color_manual(values = fill_col) +

  ## Labels: USER EDITS as desired
  labs(color = "Sample Type") +
  ggtitle("NMDS example") +
  
  ## adding stress value to plot (user edits x and y to desired location)
  annotate(geom = "label", x = 0.7, y = 1.8, 
           label = sprintf("Stress: %.6f", NMDS$stress), hjust = 0, vjust = 1, 
           label.size = NA, fontface = "italic", color = "grey30", size = 2.75, fill="white") +
  
  ## Theme: USER EDITS as desired
  theme_bw() +
  theme(
    legend.title = element_text(face = "bold", size=12),
    legend.position = "right",
    axis.title.y = element_text(margin = margin(t = 0, r = 10, b = 0, l = 0), size=10, face="bold"),
    axis.title.x = element_text(margin = margin(t = 10, r = 0, b = 0, l = 0), size=10, face="bold"),
  ) 

## USER EDITS WIDTH AND HEIGHT TO DESIRED   
ggsave("docs/eDNA 12S metab/example_output/figures/NMDS_phyloseq.png", width = 8, height = 6)
```


## Statistics 

Test: PERMANOVA 
- * indicates an interaction (SampleType*Month). Usually we are interested in the interaction of our factors.  
- + indicates an additive effect (SampleType + Month)

The output will tell you which factors significantly impact the community assemblage (matrix). 

```{r}
adonis2(bray_df ~ Area, data = sample_df, permutations = 99)
```

Additional test: Pairwise PERMANOVA 

The output will tell you which specific variable within each factor is driving the significant effects. 

```{r}
pairwise.adonis2(bray_df ~ Area, data = sample_df)
```
